fix: keep functional structure baseline clean
